refactor thor run commands
