+++ lib/tasks/run.thor
@@ -1,5 +1,5 @@
-require_relative 'runnable'
+# require_relative 'runnable'
 
-class Run < Runnable
+# class Run < Runnable
 
-end
+# end
