+++ lib/tasks/runnable.rb
@@ -17,28 +17,28 @@ class Runnable < Thor
     Kiba::Extend.config.job.verbosity = options[:verbosity].to_sym
   end
 
-  def resolve_job(key)
-    Kiba::Extend.registry.resolve(key)
-  rescue Dry::Container::KeyError
-    puts "No job with key: #{key}"
-    :failure
-  end
+  # def resolve_job(key)
+  #   Kiba::Extend.registry.resolve(key)
+  # rescue Dry::Container::KeyError
+  #   puts "No job with key: #{key}"
+  #   :failure
+  # end
 
-  def resolve_creator(job)
-    creator = job.creator
-    return creator if creator
+  # def resolve_creator(job)
+  #   creator = job.creator
+  #   return creator if creator
 
-    puts "No creator method for #{job.key}"
-    :failure
-  end
+  #   puts "No creator method for #{job.key}"
+  #   :failure
+  # end
 
-  def run_job(key)
-    job = resolve_job(key)
-    return if job == :failure
+  # def run_job(key)
+  #   job = resolve_job(key)
+  #   return if job == :failure
 
-    creator = resolve_creator(job)
-    return if creator == :failure
+  #   creator = resolve_creator(job)
+  #   return if creator == :failure
 
-    creator.call
-  end
+  #   creator.call
+  # end
 end
